Add check to prevent recursing in "CMakeFiles" directories
Added `NOT ${subdir} MATCHES "CMakeFiles"` to CMakeLists.txt in root,
exercises and libraries directories because they were causing `cmake .`
to fail building on both my windows11 and macOS Sonoma (intel) running zsh.
As far as I can tell this can do no harm and potentially be of help to
others using this setup in the future.